Move markdown rendered to composable
