+++ client/src/components/Workflow/Invocation/Export/InvocationExportPluginCard.vue
@@ -1,27 +1,12 @@
 <script setup>
 import { computed } from "vue";
 import { BCard, BCardTitle, BButtonToolbar, BButtonGroup } from "bootstrap-vue";
-import MarkdownIt from "markdown-it";
 import { InvocationExportPlugin } from "./model";
 import StsDownloadButton from "components/StsDownloadButton.vue";
 import ExportToRemoteButton from "components/Workflow/Invocation/Export/ExportToRemoteButton.vue";
+import { useMarkdown } from "composables/useMarkdown";
 
-//TODO: Move to own component
-const md = MarkdownIt();
-var defaultRender =
-    md.renderer.rules.link_open ||
-    function (tokens, idx, options, env, self) {
-        return self.renderToken(tokens, idx, options);
-    };
-md.renderer.rules.link_open = function (tokens, idx, options, env, self) {
-    var aIndex = tokens[idx].attrIndex("target");
-    if (aIndex < 0) {
-        tokens[idx].attrPush(["target", "_blank"]);
-    } else {
-        tokens[idx].attrs[aIndex][1] = "_blank";
-    }
-    return defaultRender(tokens, idx, options, env, self);
-};
+const md = useMarkdown({ openLinksInNewPage: true });
 
 const props = defineProps({
     exportPlugin: { type: InvocationExportPlugin, required: true },
